Interpolate when highways exit and reenter map boundaries
Cropping was leaving deleting the nodes, but leaving their keys in
highway node lists, which was breaking anything that needed
`node[highway.node[x]]`.

Tests files were also isolated in modules as the added tests behaved
differently when the file was run in isolation, and other existing
test files couldn't be run in isolation.